requirements.txt
beautifulsoup4
lxml
pandas
python-telegram-bot
